third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/sandboxed-iframe-navigator-serviceworker.html

<!DOCTYPE html>
<title>Accessing navigator.serviceWorker in sandboxed iframe.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>
<body>
<script>
var lastCallbackId = 0;
var callbacks = {};
function postMessageAndWaitResult(frame) {
  return new Promise(function(resolve) {
    var id = ++lastCallbackId;
    callbacks[id] = resolve;
    frame.contentWindow.postMessage({id:id}, '*');
  });
}

window.onmessage = function (e) {
  message = e.data;
  var id = message['id'];
  var calback = callbacks[id];
  delete callbacks[id];
  calback(message['result']);
};

promise_test(function(t) {
    var url = 'resources/sandboxed-iframe-navigator-serviceworker-iframe.html';
    var frame;
    return with_iframe(url)
      .then(function(f) {
          frame = f;
          return postMessageAndWaitResult(f);
        })
      .then(function(result) {
          frame.remove();
          assert_equals(result, 'ok');
          t.done();
        });
  }, 'Accessing navigator.serviceWorker in normal iframe should not throw.');

promise_test(function(t) {
    var url = 'resources/sandboxed-iframe-navigator-serviceworker-iframe.html';
    var frame;
    return with_sandboxed_iframe(url, 'allow-scripts')
      .then(function(f) {
          frame = f;
          return postMessageAndWaitResult(f);
        })
      .then(function(result) {
          frame.remove();
          assert_equals(
              result,
              'SecurityError: Failed to read the \'serviceWorker\' property from \'Navigator\': Service worker is disabled because the context is sandboxed and lacks the \'allow-same-origin\' flag.');
          t.done();
        });
  }, 'Accessing navigator.serviceWorker in sandboxed iframe should throw.');

promise_test(function(t) {
    var url = 'resources/sandboxed-iframe-navigator-serviceworker-iframe.html';
    var frame;
    return with_sandboxed_iframe(url, 'allow-scripts allow-same-origin')
      .then(function(f) {
          frame = f;
          return postMessageAndWaitResult(f);
        })
      .then(function(result) {
          frame.remove();
          assert_equals(result, 'ok');
          t.done();
        });
  },
  'Accessing navigator.serviceWorker in sandboxed iframe with ' +
  'allow-same-origin flag should not throw.');

promise_test(function(t) {
    var url = 'resources/sandboxed-iframe-navigator-serviceworker-iframe.html';
    var frame;
    return new Promise(function(resolve) {
          frame = document.createElement('iframe');
          frame.sandbox = '';
          frame.src = url;
          frame.onload = resolve;
          document.body.appendChild(frame);
          // Switching the sandbox attribute during loading the iframe.
          frame.sandbox = 'allow-scripts allow-same-origin';
        })
      .then(function() {
          return postMessageAndWaitResult(frame)
        })
      .then(function(result) {
          frame.remove();
          // TODO(nhiroki): According to the HTML5 spec, dynamically changing
          // the sandbox attribute does not have to immediately affect the
          // restriction, so in this case, the frame would be expected to act as
          // if it still doesn't have 'allow-scripts allow-same-origin' and to
          // raise a security exception when 'serviceWorker' attribute is
          // accessed.
          // https://html.spec.whatwg.org/multipage/embedded-content.html#attr-iframe-sandbox
          assert_equals(
              result,
              'InvalidStateError: Failed to get a ServiceWorkerRegistration: The document is in an invalid state.');
          t.done();
        });
  }, 'Switching iframe sandbox attribute during loading the iframe should ' +
     'not crash (http://crbug.com/532855)');

</script>
</body>
